+++ RR-Common/src/main/java/com/scor/rr/repository/AdjustmentBasisRepository.java
@@ -3,5 +3,5 @@ package com.scor.rr.repository;
 import com.scor.rr.domain.AdjustmentBasis;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdjustmentBasisRepository extends JpaRepository<AdjustmentBasis, Long> {
+public interface AdjustmentBasisRepository extends JpaRepository<AdjustmentBasis, Integer> {
 }
