+++ RR-Common/src/main/java/com/scor/rr/repository/AdjustmentReturnPeriodBandingParameterRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 
-public interface AdjustmentReturnPeriodBandingParameterRepository extends JpaRepository<ReturnPeriodBandingAdjustmentParameter,Integer> {
+public interface AdjustmentReturnPeriodBandingParameterRepository extends JpaRepository<ReturnPeriodBandingAdjustmentParameter, Integer> {
 //    @Query("select p from AdjustmentReturnPeriodBandingParameterEntity p inner join AdjustmentNodeEntity n where p.adjustmentNode = n and n.adjustmentNodeId = :id")
-    List<ReturnPeriodBandingAdjustmentParameter> findByAdjustmentNodeAdjustmentNodeId(Long id);
+    List<ReturnPeriodBandingAdjustmentParameter> findByAdjustmentNodeAdjustmentNodeId(Integer id);
 
-    void deleteByAdjustmentNode_AdjustmentNodeId(Long adjustmentNode_adjustmentNodeId); // TODO framework ?
+    void deleteByAdjustmentNode_AdjustmentNodeId(Integer adjustmentNode_adjustmentNodeId); // TODO framework ?
 }
