+++ RR-Common/src/main/java/com/scor/rr/repository/AdjustmentScalingParameterRepository.java
@@ -3,9 +3,9 @@ package com.scor.rr.repository;
 import com.scor.rr.domain.ScalingAdjustmentParameter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdjustmentScalingParameterRepository extends JpaRepository<ScalingAdjustmentParameter, Long> {
+public interface AdjustmentScalingParameterRepository extends JpaRepository<ScalingAdjustmentParameter, Integer> {
 //    @Query("select p from AdjustmentScalingParameterEntity p inner join AdjustmentNodeEntity n on p.adjustmentNode = n where n.adjustmentNodeId = :nodeId")
-    ScalingAdjustmentParameter findByAdjustmentNodeAdjustmentNodeId(Long nodeId);
+    ScalingAdjustmentParameter findByAdjustmentNodeAdjustmentNodeId(Integer nodeId);
 
-    void deleteByAdjustmentNode_AdjustmentNodeId(Long adjustmentNodeId);
+    void deleteByAdjustmentNode_AdjustmentNodeId(Integer adjustmentNodeId);
 }
