+++ RR-Common/src/main/java/com/scor/rr/repository/AdjustmentThreadRepository.java
@@ -3,6 +3,6 @@ package com.scor.rr.repository;
 import com.scor.rr.domain.AdjustmentThreadEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdjustmentThreadRepository extends JpaRepository<AdjustmentThreadEntity, Long> {
-    AdjustmentThreadEntity getAdjustmentThreadEntityByFinalPLT_PltHeaderId(Long pltHeaderId);
+public interface AdjustmentThreadRepository extends JpaRepository<AdjustmentThreadEntity, Integer> {
+    AdjustmentThreadEntity getAdjustmentThreadEntityByFinalPLT_PltHeaderId(Integer pltHeaderId);
 }
